Trait tuple_utils::Append

pub trait Append<T> {
    type Output;
    fn append(self, _: T) -> Self::Output;
}

Helper trait to allow Appending of tuples

Associated Types

type Output

Loading content...
Required methods

fn append(self, _: T) -> Self::Output

Append T onto the end of the tuple returning a new tuple with the existing elements and T

Loading content...