Trait tuple_utils::Prepend

pub trait Prepend<T> {
    type Output;
    fn prepend(self, _: T) -> Self::Output;
}

Helper trait to allow Perpending of tuples

Associated Types

type Output

Required Methods

fn prepend(self, _: T) -> Self::Output

Append T onto the start of the tuple returning a new tuple with all the elements from shifted over one row and T in the first slot