Trait tuple_utils::Split

pub trait Split {
    type Left;
    type Right;
    fn split(self) -> (Self::Left, Self::Right);
}

Tries to split a tuple into two tuples if the tuple is odd sized the Right side will contain the extra element

Associated Types

type Left

type Right

Required Methods

fn split(self) -> (Self::Left, Self::Right)