Trait tuple_utils::Merge

pub trait Merge<T> {
    type Output;
    fn merge(self, _: T) -> Self::Output;
}

Helper trait that allow for merging of tuples

Associated Types

type Output

Required Methods

fn merge(self, _: T) -> Self::Output

merge LHS with RHS returning a new tuple that contains the elements of both tuples ordering is preserved such that LHS elements come before RHS elements.